match(itbox): itBox_UnkMotion0_Anim, UnkMotion0_Coll, UnkMotion4_Phys
Matches 3 functions:
- itBox_UnkMotion0_Anim: life timer check with it_802787B4 callback
- itBox_UnkMotion0_Coll: collision callback with it_8026D62C pattern
- itBox_UnkMotion4_Phys: fall physics with it_80272860 and it_80274658

Also extends itDosei_ItemVars struct with xDF0, xDF4, xDF8 fields.

+++ src/melee/it/itCommonItems.h
@@ -111,8 +111,8 @@ typedef struct itDosei_ItemVars {
     f32 xDDC;
     s32 xDE0;
     Vec3 xDE4;
-    f32 xDF0;
-    f32 xDF4;
+    s32 xDF0;
+    s32 xDF4;
     Vec3 xDF8;
 } itDosei_ItemVars;
